try fixing ptr

+++ printptr.c
@@ -2,10 +2,10 @@
 
 void printptr(void *n, int *ReturnValue)
 {
-	unsigned int *arg1;
+	long unsigned int *arg1;
 	char arg2;
 
-	arg1 = (unsigned int *) n;
+	arg1 = (long unsigned int *) n;
 	arg2 = '0';
 	printchar(&arg2, ReturnValue);
 	arg2 = 'x';
